pyproject.toml
[project]
name = "olist-mart"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Santos",email = "oliveirasgah@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb (==1.10.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
